requirements.txt
cachelib==0.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.2
Flask-Session==0.4.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
